+++ src/tm_devices/drivers/mainframes/mp5xxx/mp5xxx.py
@@ -139,7 +139,10 @@ class MP5xxx(CommonTSPErrorCheckMixin, TSPControl, Mainframe, ABC):
             raise ValueError(msg)
 
         try:
-            model_returned = self.query(f"print(slot[{slot}].model)")
+            model_returned = self.query(
+                f"if slot[{slot}] == nil then print('') else print(slot[{slot}].model) end",
+                allow_empty=True,
+            )
             if model_returned.find("-") == -1:
                 msg = f"{model_returned} Module in slot {slot} is an invalid module."
                 raise ValueError(msg)  # noqa: TRY301
@@ -150,11 +153,10 @@ class MP5xxx(CommonTSPErrorCheckMixin, TSPControl, Mainframe, ABC):
 
             supported_modules = ", ".join(module_alias)
 
-            installed_module = self.query(f"print(slot[{slot}].model)")
             msg = (
                 f"No supported {module_type} module found in slot {slot}. "
                 f"The supported {module_type} modules are: {supported_modules}. "
-                f"Currently installed module: {installed_module}."
+                f"Currently installed module: {model_returned}."
             )
             raise TypeError(msg)
         except ValueError as error: